# server/src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Connection Manager API"
  version: "1.6.0"
  title: "Connection Manager API"
  contact:
    email: "ramiro@modusbox.com"
host: "localhost:3001"
basePath: "/api"
...
schemes:
- "http"
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: "password"
     # The flow used by the OAuth2 security scheme. Valid values are "implicit", "password", "application" or "accessCode".
    tokenUrl: https://WSO2_IM_SERVERl:9443/oauth2/token
    scopes:
      pta: Portal Technical Administrator
      mta: Mojabox Technical Administrator
      everyone: All logged in users
security:
  - OAuth2: [ pta ] # most restrictive role
paths:
...
  /environments/{envId}/hub/endpoints:
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Returns all Hub endpoints"
      security:
        - OAuth2: [ everyone ]
      operationId: "getHubEndpoints"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "Environment id"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
  /environments/{envId}/hub/endpoints/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ everyone ]
      operationId: "getHubEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    put:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Update an endpoint entry"
      operationId: "updateHubEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/EndPointAny"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    delete:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      operationId: "deleteHubEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "HubNetworkConfig"      
  /environments/{envId}/hub/endpoints/ingress/ips:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    post:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Adds a new IP entry to the Hub Ingress endpoint"
      description: "Adds the entry with state NEW"
      operationId: "createHubIngressIp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Hub ingress IP entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "HubNetworkConfig"      
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get the Hub Ingress IPs"
      security:
        - OAuth2: [ everyone ]
      description: ""
      operationId: "getHubIngressIps"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "HubNetworkConfig"   
  /environments/{envId}/hub/endpoints/ingress/ips/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ everyone ]
      operationId: "getHubIngressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    put:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Update an endpoint entry"
      operationId: "updateHubIngressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    delete:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      operationId: "deleteHubIngressIpEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "HubNetworkConfig"      
  /environments/{envId}/hub/endpoints/ingress/urls:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    post:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Adds a Hub Ingress URL"
      description: "Adds a Hub ingress URL with state NEW"
      operationId: "createHubIngressUrl"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Hub ingress URL"
        required: true
        schema:
          $ref: "#/definitions/InputURLValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointURL"
      x-swagger-router-controller: "HubNetworkConfig"   
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get the Hub Ingress URLs"
      security:
        - OAuth2: [ everyone ]
      description: ""
      operationId: "getHubIngressUrls"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointURL"
      x-swagger-router-controller: "HubNetworkConfig"   
  /environments/{envId}/hub/endpoints/ingress/urls/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ everyone ]
      operationId: "getHubIngressUrlEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    put:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Update an endpoint entry"
      operationId: "updateHubIngressUrlEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputURLValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    delete:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      operationId: "deleteHubIngressUrlEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "HubNetworkConfig"      
  /environments/{envId}/hub/endpoints/egress/ips:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    post:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Adds a new IP entry to the Hub Egress endpoint"
      description: "Adds the entry with state NEW"
      operationId: "createHubEgressIp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Hub egress IP entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "HubNetworkConfig"      
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get the Hub Egress IPs"
      security:
        - OAuth2: [ everyone ]
      description: ""
      operationId: "getHubEgressIps"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "HubNetworkConfig"   
  /environments/{envId}/hub/endpoints/egress/ips/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ everyone ]
      operationId: "getHubEgressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    put:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Update an endpoint entry"
      operationId: "updateHubEgressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "HubNetworkConfig"      
    delete:
      tags:
      - "Hub-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      operationId: "deleteHubEgressIpEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "HubNetworkConfig"      
...
definitions:
...
  URLEntry:
    properties:
      url:
        type: "string"
        description: "URL"
  IPEntry:
    properties:
      address:
        type: "string"
        description: "A single IP address or a set of address using CIDR notation \
        \ IP:\
        \ a.b.c.d\
        \ where 0 <= [a,b,c,d] <= 255\
        \ CIDR:\
        \ a.b.c.d/e\
        \ where 0 <= [a,b,c,d] <= 255, and 0 <= e <= 32"
        pattern: '^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(\/([0-9]|[1-2][0-9]|3[0-2]))?$'
        example:
          "163.10.24.28/30"
      ports:
        type: "array"
        description: "List of ports or port ranges"
        items:
          type: "string"
          description: "Individual port or port range"
          example: "8000-8080"
        example:
          - "80"
          - "8000-8080"
  BaseEndpoint:
    description: Common properties to both IP and URL endpoints
    properties:
      id:
        type: "integer"
        format: "int32"
      type:
        type: "string"
        enum:
        - "IP"
        - "URL"
      direction:
        type: "string"
        enum:
        - "INGRESS"
        - "EGRESS"
      state:
        type: "string"
        enum:
        - "NEW"
        - "CONFIRMED"
        - "INVALID"
      createdDate:
        type: "string"
        format: "date-time"
      createdBy:
        type: "string"
      dfspId:
        type: "string"
  EndPointIp:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            $ref: "#/definitions/IPEntry"
  EndPointURL:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            $ref: "#/definitions/URLEntry"
  EndPointAny:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            description: It's *either* {url} OR {address and ports}. 
            allOf:
              - $ref: "#/definitions/IPEntry"
              - $ref: "#/definitions/URLEntry"
  InputIPValue:
    type: object
    properties:
      value:
        $ref: "#/definitions/IPEntry"
  InputURLValue:
    type: object
    properties:
      value:
        $ref: "#/definitions/URLEntry"
...